# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "An API to build, play and teach bricks"
  version: "1.0.0"
  title: "Brillder"
  termsOfService: ""
  contact:
    email: "lindsaymacvean@scholar6.org"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "swagger.scholar6.org"
...
schemes:
- "https"
- "http"
paths:
...
  /brick/cover:
    post:
      tags:
      - "brick"
      summary: "add brick cover image with description"
      operationId: "addBrickCover"
      parameters:
      - in: "body"
        name: "body"
        description: "Brick Cover Image Model"
        required: true
        schema:
          $ref: "#/definitions/CoverImageModel"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      x-swagger-router-controller: "Brick"
  /brick/setCoreLibrary/{brickId}/{isCore}:
    put:
      tags:
      - "brick"
      summary: "Sets whether a brick is in the core library or not."
      operationId: "brickSetCoreLibraryBrickIdIsCorePUT"
      parameters:
      - name: "brickId"
        in: "path"
        required: true
        type: "integer"
      - name: "isCore"
        in: "path"
        required: true
        type: "boolean"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
  /brick/assignEditor:
    put:
      tags:
      - "brick"
      summary: "Assign editors to brick"
      operationId: "assignEditor"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_5"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
  /brick/assignCoAuthor:
    put:
      tags:
      - "brick"
      summary: "Assign co-authors to brick"
      operationId: "assignCoAuthor"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/body_6"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
...
  /bricks/search:
    post:
      tags:
      - "brick"
      summary: "Search over all bricks"
      operationId: "bricksSearchPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Search model"
        required: true
        schema:
          $ref: "#/definitions/SearchModel"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/ArrayOfBricks"
      x-swagger-router-controller: "Brick"
  /bricks/search/public:
    post:
      tags:
      - "brick"
      summary: "Search over all bricks"
      operationId: "bricksPublicSearchPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Search model"
        required: true
        schema:
          $ref: "#/definitions/SearchModel"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/ArrayOfBricks"
      x-swagger-router-controller: "Brick"
  /bricks/byStatus/{status}:
    get:
      tags:
      - "brick"
      summary: "Get all bricks"
      operationId: "bricksByStatusStatusGET"
      parameters:
      - name: "status"
        in: "path"
        required: true
        type: "integer"
        enum:
        - 1
        - 2
        - 3
        - 4
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/ArrayOfBricks"
      x-swagger-router-controller: "Brick"
...
  /bricks/comment:
    post:
      tags:
      - "brick"
      - "comment"
      summary: "Post a new comment on a brick."
      operationId: "bricksCommentPOST"
      parameters:
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/body_3"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Comment"
      x-swagger-router-controller: "Brick"
  /bricks/comments/{brickId}:
    get:
      tags:
      - "brick"
      - "comment"
      summary: "Get all comments for a brick."
      operationId: "bricksCommentsBrickIdGET"
      parameters:
      - name: "brickId"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Comment"
      x-swagger-router-controller: "Brick"
...
  /brick/{id}:
    get:
      tags:
      - "brick"
      summary: "Get specific brick by id"
      operationId: "brickIdGET"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      x-swagger-router-controller: "Brick"
    delete:
      tags:
      - "brick"
      summary: "Delete specific brick by id"
      operationId: "brickIdDELETE"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
  /brick/build/{id}:
    post:
      tags:
      - "brick"
      summary: "Set brick to build status"
      description: ""
      operationId: "brickBuildIdPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      security:
      - brillder_auth:
        - "write:bricks"
        - "read:bricks"
      x-swagger-router-controller: "Brick"
  /brick/review/{id}:
    post:
      tags:
      - "brick"
      summary: "Set brick to review status"
      description: ""
      operationId: "brickReviewIdPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      security:
      - brillder_auth:
        - "write:bricks"
        - "read:bricks"
      x-swagger-router-controller: "Brick"
...
  /brick/publish/{id}:
    post:
      tags:
      - "brick"
      summary: "Set brick to publish status"
      description: ""
      operationId: "brickPublishIdPOST"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      security:
      - brillder_auth:
        - "write:bricks"
        - "read:bricks"
      x-swagger-router-controller: "Brick"
  /brick:
    post:
      tags:
      - "brick"
      summary: "Add a new brick"
      description: ""
      operationId: "addBrick"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Brick object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/Brick"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      security:
      - brillder_auth:
        - "write:bricks"
        - "read:bricks"
      x-swagger-router-controller: "Brick"
    put:
      tags:
      - "brick"
      summary: "Update an existing brick"
      description: ""
      operationId: "updateBrick"
      consumes:
      - "application/json"
      - "application/xml"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Brick object that needs to be updated"
        required: true
        schema:
          $ref: "#/definitions/Brick"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Brick"
      security:
      - brillder_auth:
        - "write:bricks"
        - "read:bricks"
      x-swagger-router-controller: "Brick"
  /brick/{brickId}/comment/{commentId}:
    get:
      tags:
      - "brick"
      - "comment"
      summary: "Get a specific comment on a brick."
      operationId: "brickBrickIdCommentCommentIdGET"
      parameters:
      - name: "brickId"
        in: "path"
        required: true
        type: "integer"
      - name: "commentId"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
          schema:
            $ref: "#/definitions/Comment"
      x-swagger-router-controller: "Brick"
  /brick/assignment/{assignmentId}/unarchive:
    post:
      tags:
      - "brick"
      summary: "Archive assignment"
      description: ""
      operationId: "archiveAssignment"
      parameters:
      - name: "assignmentId"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
...
  /brick/assignment/{assignmentId}/archive:
    post:
      tags:
      - "brick"
      summary: "Archive assignment"
      description: ""
      operationId: "archiveAssignment"
      parameters:
      - name: "assignmentId"
        in: "path"
        required: true
        type: "integer"
      responses:
        "200":
          description: "ok"
      x-swagger-router-controller: "Brick"
  /brick/assignStudents/{brickId}:
    post:
      tags:
      - "brick"
      summary: "Assign students to brick"
      description: ""
      operationId: "assignStudents"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "brickId"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "studentsIds"
        required: false
        schema:
          $ref: "#/definitions/studentsIds"
      responses:
        "200":
          description: "ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Assignment"
      x-swagger-router-controller: "Brick"
  /brick/assignClasses/{brickId}:
    post:
      tags:
      - "brick"
      summary: "Assign classes to brick"
      description: ""
      operationId: "assignClasses"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "brickId"
        in: "path"
        required: true
        type: "integer"
      - in: "body"
        name: "classesIds"
        required: true
        schema:
          $ref: "#/definitions/classesIds"
      responses:
        "200":
          description: "ok"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Assignment"
      x-swagger-router-controller: "Brick"
...
securityDefinitions:
  brillder_auth:
    type: "oauth2"
    authorizationUrl: "https://example.com/oauth/authorize"
    flow: "implicit"
    scopes:
      write:bricks: "modify bricks in your account"
      read:bricks: "read your bricks"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  ArrayOfBricks:
    type: "array"
    items:
      $ref: "#/definitions/Brick"
...
  ArrayOfSubjects:
    type: "array"
    items:
      $ref: "#/definitions/Subject"
...
  Question:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      order:
        type: "integer"
        format: "int64"
      contentBlocks:
        type: "string"
      questionType:
        type: "string"
        enum:
        - "None"
        - "ShortAnswer"
        - "ChooseOne"
        - "ChooseSeveral"
        - "VerticalShuffle"
        - "HorizontalShuffle"
        - "PairMatch"
        - "Categorise"
        - "MissingWord"
        - "WordHighlighting"
        - "LineHighlighting"
    xml:
      name: "Question"
    example:
      contentBlocks: "contentBlocks"
      id: 7
      questionType: "None"
      order: 9
  Brick:
    type: "object"
    required:
    - "questions"
    - "subject"
    - "title"
    properties:
      id:
        type: "integer"
        format: "int64"
      attemptsCount:
        type: "integer"
        format: "int64"
      brickLength:
        type: "string"
        enum:
        - "0"
        - "20"
        - "40"
        - "60"
      subjectId:
        type: "integer"
      subject:
        $ref: "#/definitions/Subject"
      locked:
        type: "boolean"
      isCore:
        type: "boolean"
      alternativeSubject:
        type: "string"
      academicLevel:
        type: "number"
      title:
        type: "string"
      brief:
        type: "string"
      sponsorLogo:
        type: "string"
      sponsorUrl:
        type: "string"
      prep:
        type: "string"
      synthesis:
        type: "string"
      revisionLog:
        type: "string"
      openQuestion:
        type: "string"
      created:
        type: "string"
        format: "date-time"
      updated:
        type: "string"
        format: "date-time"
      questions:
        type: "array"
        items:
          $ref: "#/definitions/Question"
      keywords:
        type: "array"
        items:
          $ref: "#/definitions/Keyword"
    xml:
      name: "Brick"
    example:
      alternativeSubject: "alternativeSubject"
      academicLevel: 1
      brief: "brief"
      synthesis: "synthesis"
      openQuestion: "openQuestion"
      subject:
        color: "color"
        name: "name"
        checked: true
        id: 5
        bricksCount: 5
        publishedBricksCount: 2
      created: "2000-01-23T04:56:07.000+00:00"
      keywords:
      - name: "batterfly"
        id: 1
      - name: "new"
      questions:
      - contentBlocks: "contentBlocks"
        id: 7
        questionType: "None"
        order: 9
      - contentBlocks: "contentBlocks"
        id: 7
        questionType: "None"
        order: 9
      revisionLog: "revisionLog"
      title: "title"
      subjectId: 1
      attemptsCount: 6
      brickLength: "0"
      prep: "prep"
      id: 0
      locked: true
      updated: "2000-01-23T04:56:07.000+00:00"
...
  Comment:
    type: "object"
    required:
    - "author"
    - "brick"
    - "id"
    - "readBy"
    - "text"
    - "timestamp"
    properties:
      id:
        type: "number"
      text:
        type: "string"
      brick:
        $ref: "#/definitions/Brick"
      question:
        $ref: "#/definitions/Question"
      author:
        $ref: "#/definitions/UserShort"
      parent:
        $ref: "#/definitions/Comment"
      children:
        type: "array"
        items:
          $ref: "#/definitions/Comment"
      timestamp:
        type: "string"
        format: "date-time"
      readBy:
        type: "array"
        items:
          $ref: "#/definitions/Comment_readBy"
    example:
      id: 0
      text: "Comment text..."
      brick:
        alternativeSubject: "alternativeSubject"
        brief: "brief"
        synthesis: "synthesis"
        openQuestion: "openQuestion"
        created: "2000-01-23T04:56:07.000+00:00"
        revisionLog: "revisionLog"
        title: "title"
        subjectId: 1
        attemptsCount: 6
        brickLength: "0"
        prep: "prep"
        id: 0
        locked: true
        updated: "2000-01-23T04:56:07.000+00:00"
      question:
        contentBlocks: "contentBlocks"
        id: 7
        questionType: "None"
        order: 9
      author:
        firstName: "firstName"
        lastName: "lastName"
        subjects:
        - ""
        - ""
        roles:
        - roleId: 2
        - roleId: 2
        id: 6
        tutorialPassed: true
        email: "email"
      children: []
      timestamp: "2000-01-23T04:56:07.000+00:00"
      readBy:
      - id: 1
...
  Subject:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      group:
        type: "integer"
        description: "Subject group:\n  * `0` - None\n  * `1` - Arts\n  * `2` - Languages\n\
          \  * `3` - HumanitiesAndSocialSciences\n  * `4` - GeneralTopical\n  * `5`\
          \ - Science\n  * `6` - MathsAndComputing\n"
      color:
        type: "string"
      checked:
        type: "boolean"
      name:
        type: "string"
      bricksCount:
        type: "integer"
        format: "int32"
      publishedBricksCount:
        type: "integer"
        format: "int32"
    xml:
      name: "Subject"
    example:
      color: "color"
      name: "name"
      checked: true
      group: 1
      id: 5
      bricksCount: 5
      publishedBricksCount: 2
  Keyword:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
    xml:
      name: "Keyword"
    example:
      name: "butterfly"
      id: 5
...
  UserShort:
    type: "object"
    properties:
      subjects:
        type: "array"
        items:
          $ref: "#/definitions/ArrayOfSubjects"
      email:
        type: "string"
      firstName:
        type: "string"
      id:
        type: "integer"
        format: "int64"
      lastName:
        type: "string"
      tutorialPassed:
        type: "boolean"
      roles:
        type: "array"
        items:
          $ref: "#/definitions/User_roles"
    xml:
      name: "User"
    example:
      firstName: "firstName"
      lastName: "lastName"
      subjects:
      - ""
      - ""
      roles:
      - roleId: 2
      - roleId: 2
      id: 6
      tutorialPassed: true
      email: "email"
...
  CoverImageModel:
    type: "object"
    properties:
      brickId:
        type: "integer"
      coverImage:
        type: "string"
      coverImageSource:
        type: "string"
      coverImageCaption:
        type: "string"
  SearchModel:
    type: "object"
    properties:
      searchString:
        type: "string"
    example:
      searchString: "searchString"
...
  User_roles:
    type: "object"
    properties:
      roleId:
        type: "integer"
    example:
      roleId: 2
...
  body_3:
    type: "object"
    properties:
      text:
        type: "string"
      brickId:
        type: "integer"
      questionId:
        type: "integer"
      parentId:
        type: "integer"
  Comment_readBy:
    type: "object"
    properties:
      id:
        type: "number"
...
  body_5:
    type: "object"
    required:
    - "brickId"
    - "editorIds"
    properties:
      brickId:
        type: "integer"
      editorIds:
        type: "array"
        items:
          type: "integer"
  body_6:
    type: "object"
    required:
    - "brickId"
    - "coAuthors"
    properties:
      brickId:
        type: "integer"
      coAuthors:
        type: "array"
        items:
          type: "integer"
  studentsIds:
    type: "object"
    required:
    - "studentsIds"
    properties:
      studentsIds:
        type: "array"
        items:
          type: "integer"
      deadline:
        type: "string"
        format: "date-time"
  classesIds:
    type: "object"
    required:
    - "classesIds"
    properties:
      classesIds:
        type: "array"
        items:
          type: "integer"
      deadline:
        type: "string"
        format: "date-time"
...